third_party/WebKit/LayoutTests/inspector/sources/debugger/async-operation-events-expected.txt

Tests debugger AsyncOperation events while in a debugger stepping session.

Set timer for test function.
Executing StepOver...
Executing StepOver...
==> AsyncOperationStarted: setTimeout
Executing StepInto...
==> AsyncOperationCompleted
Call stack:
    0) callback1 (async-operation-events.html:13)
    [setTimeout]
    0) testFunction (async-operation-events.html:10)

Executing StepOver...
Executing StepOver...
==> AsyncOperationStarted: setTimeout
Executing Resume...
==> AsyncOperationsCleared
Executing Resume...
